Split macOS input backend
